Broken: Working on Entry table

+++ WHY_NOT/src/dao/interfaces/DatabaseDAO.java
@@ -7,15 +7,15 @@ import model.EntryFile;
 import dao.hibernate.GenericDAO;
 
 public interface DatabaseDAO extends GenericDAO<Database, String> {
-	int getValidCount(Database db);
+	long getValidCount(Database db);
 
 	Set<EntryFile> getValidEntries(Database db);
 
-	int getMissingCount(Database db);
+	long getMissingCount(Database db);
 
 	Set<EntryFile> getMissingEntries(Database db);
 
-	int getObsoleteCount(Database db);
+	long getObsoleteCount(Database db);
 
 	Set<EntryFile> getObsoleteEntries(Database db);
